Shifting test answer area into separate component.
The component, when inflection tables are implemented, may either show a
textarea element, or an inflection table.
