+++ src/app/tests/[sheet]/components/test-question/test-question.module.css
@@ -3,16 +3,6 @@
     padding-block: 10px;
 }
 
-.answerbox {
-    background-color: rgb(var(--colour-ash));
-    color: rgb(var(--colour-pearl));
-    resize: none;
-    font-size: 40px;
-    font-family: inherit;
-    width: 800px;
-    height: 200px;
-}
-
 .starbox {
     padding-block: 10px;
 }
